Converts more Classes into Typescript
- Converts AndroidSdkTools
- Converts GradleWrapper
- Converts AndroidSdkToolsSpec
- Updates other parts of the code to be compatible with the updated
  ones
